[maps] organize helper functions and add tests (#1377)
* [maps] organize helper functions and add tests

* [maps] update test locations
